+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 The `send_array` and `retrieve_array` methods can now handle 3D arrays
 as well as 2D ones.
 
+The ds9 connection has now gained a debug field which can be set to
+`True` to display the SAMP return value for each `set` or `get` call.
+
 ## Version 0.0.5 - 2025-01-27
 
 Added the `retrieve_array` method, which saves the current frame to a