+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## Version 0.0.6 - 2025-01-28
 
 The `send_array` and `retrieve_array` methods can now handle 3D arrays
-as well as 2D ones.
+as well as 2D ones. Data in RGB, HLS, or HSV format is identified by
+setting the `cube` argument in `send_array` to one of: `Cube.RGB`,
+`Cube.HLS`, or `Cube.HSV`.
 
 Commands which return data via a url, such as "data", will now return
 the contents of this file, rather than returning `None` (this is only